quelques bugs fixés

+++ app/controllers/profiles_controller.ts
@@ -5,6 +5,7 @@ import { cuid } from '@adonisjs/core/helpers'
 import type { HttpContext } from '@adonisjs/core/http'
 import app from '@adonisjs/core/services/app'
 import { unlink } from 'fs/promises'
+import Helper from '../helpers/helper.js'
 @inject()
 export default class ProfilesController {
   constructor(private _userService: UserService) {}
@@ -33,14 +34,14 @@ export default class ProfilesController {
 
   async updateAvatar({ request, response, auth, session }: HttpContext) {
     if (!request.file('avatar')) {
-      session.flash('noPicture', 'You must chose a picture')
+      Helper.setFlashMessage(session, 'error', 'You must chose a picture')
 
       response.redirect().back()
     }
     const { avatar } = await request.validateUsing(profileUpdateAvatarValidator)
 
     if (avatar) {
-      if (auth.user?.avatar) {
+      if (auth.user?.avatar && auth.user?.avatar !== 'placeholder.png') {
         await unlink(app.makePath('storage/uploads/avatars', auth.user.avatar))
       }
       await avatar.move(app.makePath('storage/uploads/avatars'), {
@@ -48,7 +49,7 @@ export default class ProfilesController {
       })
       auth.user!.avatar = avatar.fileName!
       await auth.user?.save()
-
+      Helper.setFlashMessage(session, 'avatar', 'Avatar updated', undefined, false)
       response.redirect().back()
     }
   }
